sbrc2026-artifact-rag/pyproject.toml
[project]
name = "sbrc2026-artifact-rag"
version = "0.1.0"
description = "Minimal RAG artifact (Qdrant embedded + llama.cpp server) for SBRC 2026."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "openai>=2.6.0",
  "python-dotenv>=1.1.1",
  "qdrant-client==1.15.1",
  "torch>=2.0.0",
  "transformers>=4.40.0",
  "accelerate>=0.20.0",
  "llama-index-core>=0.14.5",
  "llama-index-readers-file>=0.5.4",
]

